Add customization option for `pop-to-buffer`
First of I would like to say thanks for an awesome package. I have
been thinking of creating something similar for ages as I have not
been the biggest fan of evil's jumplist and I think this is
great. Can't wait for this to be on melpa as I am not a straight user.

The default `display-buffer` action for `pop-to-buffer` with my usage
of this great package makes my skin crawl. Probably due to being an
old vim user. Personally I prefer the ACTION
`display-buffer-same-window` but obviously that is not the case for
everyone. So I thought it would be nice if the behavior could be
customized. As

```
(setq gumshoe-display-buffer-action '(display-buffer-same-window))
```

And maybe it helps someone else if the custom takes specifies the
ACTION argument to align this great package with someone else's
workflow.