separate perspective extension into separate file
This required creating a top-level file to make sure files are loaded
in the correct order. I also added an explicit flag to allow the user
to specify if they want the perspective extension automatically
configured. But I turn it on by default to not break backward
compatibility. The extension is also only loaded after perspectives of
course.