decouple backlog from entry type
This means the entry type can be changed at any point and things won't
completely break. I had to ignore errors when trying to access slots
that aren't there to make this work though. The output will be nil in
those columns though, so it should be pretty obvious anyway if someone
mistypes the schema.